samples: serial_lte_modem suppress build warnings
With NEWLIB_LIBC used, some C functions generates build warnings
Namely, toupper, tolower, isdigit...require int type parameter

+++ lib/at_cmd_parser/at_utils.h
@@ -62,7 +62,7 @@ static inline bool is_notification(char chr)
  */
 static inline bool is_valid_notification_char(char chr)
 {
-	chr = toupper(chr);
+	chr = toupper((int)chr);
 
 	if ((chr >= 'A') && (chr <= 'Z')) {
 		return true;
@@ -201,7 +201,7 @@ static inline bool is_array_stop(char chr)
  */
 static inline bool is_number(char chr)
 {
-	if (isdigit(chr) || (chr == '-') || (chr == '+')) {
+	if (isdigit((int)chr) || (chr == '-') || (chr == '+')) {
 		return true;
 	}
 
@@ -224,7 +224,7 @@ static inline bool is_command(const char *str)
 		return false;
 	}
 
-	if ((toupper(str[0]) != 'A') || (toupper(str[1]) != 'T')) {
+	if ((toupper((int)str[0]) != 'A') || (toupper((int)str[1]) != 'T')) {
 		return false;
 	}
